move TraceError, TraceResult and ExportError to opentelemetry-sdk
