+++ opentelemetry-sdk/src/trace/in_memory_exporter.rs
@@ -1,8 +1,8 @@
 use crate::error::{OTelSdkError, OTelSdkResult};
 use crate::resource::Resource;
+use crate::trace::error::{TraceError, TraceResult};
 use crate::trace::{SpanData, SpanExporter};
 use futures_util::future::BoxFuture;
-use opentelemetry::trace::{TraceError, TraceResult};
 use std::sync::{Arc, Mutex};
 
 /// An in-memory span exporter that stores span data in memory.
